挑战次数作为领取奖励条件。

+++ template/tpl_equipment.go
@@ -16,13 +16,13 @@ type TplEquipment struct {
 	DisplayScale     int32   `json:"displayScale"`
 	EquipInfoUIScale float64 `json:"equipInfoUIScale"`
 	IconRes          string  `json:"iconRes"`
-	IconScale        float64 `json:"iconScale"`
+	IconScale        int32   `json:"iconScale"`
 	ID               string  `json:"id"`
 	Name             string  `json:"name"`
 	Occupy           int32   `json:"occupy"`
 	Order            int32   `json:"order"`
 	Quality          int32   `json:"quality"`
-	StorageScale     float64 `json:"storageScale"`
+	StorageScale     int32   `json:"storageScale"`
 	TargetType       int32   `json:"targetType"`
 	Type             int32   `json:"type"`
 	UnlockUIScale    float64 `json:"unlockUIScale"`
